+++ commands/data/crn.js
@@ -5,7 +5,7 @@ const { MessageEmbed } = require("discord.js");
 module.exports = {
   name: "crn",
   description:
-    "Gives you information about any UC Davis course for this current term.",
+    "Returns CRN information on any UC Davis course for this current term.",
   hasCommand: true,
   hasSlash: true,
   hasButton: false,
@@ -82,9 +82,12 @@ const readCRNs = async (courseCode) => {
   let prevCRN = courses[0].CRN;
   let fieldDesc = "";
 
-  courses.forEach((course) => {
+  // TODO: instead of calling GetRMPData for each course, we should get instructors first, get their data, then refer to that in the for loop
+  for (const course of courses) {
     if (prevCRN !== course.CRN) {
-      fieldDesc += "Instructor: " + course.Instructor;
+      fieldDesc += `Instructor: ${
+        courses[courses.length - 1].Instructor
+      } (${await GetRMPData(courses[courses.length - 1].Instructor)})`;
       reply.addField(prevCRN, fieldDesc, true);
       prevCRN = course.CRN;
       fieldDesc = "";
@@ -103,9 +106,36 @@ const readCRNs = async (courseCode) => {
         ("0" + (course[day + "EndTime"] % 60)).slice(-2) +
         "\n";
     });
-  });
-  fieldDesc += "Instructor: " + courses[courses.length - 1].Instructor;
+  }
+  fieldDesc += `Instructor: ${
+    courses[courses.length - 1].Instructor
+  } (${await GetRMPData(courses[courses.length - 1].Instructor)})`;
   reply.addField(courses[courses.length - 1].CRN, fieldDesc, true);
 
   return { embeds: [reply] };
+};
+
+const GetRMPData = async (teacherName) => {
+  // getting the csv file and parsing it
+  const fileContent = await fs.readFile(__dirname + "/../../data/rmpData.csv");
+  const records = parse(fileContent, {
+    columns: true,
+    escape: "\\",
+    // skipLinesWithError: true,
+  });
+  const teachers = records.filter((teacher) =>
+    teacher.name
+      .toLowerCase()
+      .split(" ")
+      .slice(-1)[0]
+      .includes(
+        teacherName.toLowerCase().split(", ")[0] &&
+          teacher.name
+            .toLowerCase()
+            .split(" ")[0]
+            .includes(teacherName.toLowerCase().split(", ").slice(-1)[0])
+      )
+  );
+  if(!teachers[0] || teachers.length > 1) return "N/A";
+  return teachers[0].rating;
 };